% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changement_lettre.R
\name{remplacer_2eme_lettre}
\alias{remplacer_2eme_lettre}
\title{Remplacement de la 2e lettre}
\usage{
remplacer_2eme_lettre(chaine)
}
\arguments{
\item{chaine}{}
}
\description{
Remplacement de la 2e lettre
}
